Organize files into cli and utils
